requirements.txt
tweepy=3.5.0
RPi.GPIO=0.6.2
